+++ changelog.md
@@ -8,9 +8,11 @@ All notable changes to this project will be documented in this file.
 
 - Download 2020 year-to-date data from the Chicago Data Portal and summarize on launch
 - Add crash trends
+- Process and display of crashes by intersection
 
 ### Updated
 - Update Readme
+- Fix crashes by intersection year filtering
 
 ## [0.6] - 2020-10-11